[issue462] Remove -Wno-sign-compare and fix resulting issues.

+++ src/search/cea_heuristic.cc
@@ -395,7 +395,7 @@ void ContextEnhancedAdditiveHeuristic::initialize() {
     goal_node = &goal_problem->nodes[1];
 
     local_problem_index.resize(num_variables);
-    for (size_t var_no = 0; var_no < num_variables; ++var_no) {
+    for (int var_no = 0; var_no < num_variables; ++var_no) {
         int num_values = g_variable_domain[var_no];
         local_problem_index[var_no].resize(num_values, 0);
     }
